Rename: component/*.js -> component/*/index.js로 변경
